requirements.txt
absl-py==1.3.0
astunparse==1.6.3
black==22.10.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
easydict==1.10
flatbuffers==22.9.24
fonttools==4.37.4
gast==0.4.0
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.49.1
h5py==3.7.0
idna==3.4
imageio==2.22.2
importlib-metadata==5.0.0
imutils==0.5.4
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
labelImg==1.8.6
libclang==14.0.6
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
mypy-extensions==0.4.3
networkx==2.8.7
ninja==1.10.2.4
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post3
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.9.3
Shapely==1.8.5.post1
six==1.16.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
tifffile==2022.10.10
tomli==2.0.1
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.9.0
